Rearrange methods, extract wrapper protocol
