+++ Core/Core/Common/CommonModels/Analytics/Pendo/PendoManagerWrapper.swift
@@ -1,0 +1,32 @@
+//
+// This file is part of Canvas.
+// Copyright (C) 2026-present  Instructure, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+
+
+import Foundation
+import Pendo
+
+// This is only needed to make testing possible
+public protocol PendoManagerWrapper: AnyObject {
+    func initWith(_ url: URL)
+    func setup(_ appKey: String)
+    func startSession(_ visitorId: String?, accountId: String?, visitorData: [AnyHashable: Any]?, accountData: [AnyHashable: Any]?)
+    func endSession()
+    func track(_ event: String, properties: [AnyHashable: Any]?)
+}
+
+extension PendoManager: PendoManagerWrapper {}
